+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HFP AG: fix setup of audio connection in service level established event
 - HFP AG: fix BCM WBS setup
 - HID Device: set Report Mode as default
+- AVRCP: reject incoming second l2cap connection
 
 ### Changed
 - btstack_crypto: allow MBEDTLS config via MBEDTLS_CONFIG_FILE